Repository restructure
This commit brings revisions to the initial directory structure. Changes
includes are:
 * Move the lib package outside the tests package
 * Move aoc operations tests into a new aoc package under tests
 * Create new aoc package under lib with subpackages for each supported
   cloud. Within each cloud package, will contain packages/modules for
   various tasks (e.g. operations).
