+++ README.md
@@ -38,14 +38,14 @@ export AOC_OPS_IMAGE_REGISTRY_PASSWORD=<password>
 cd tests
 
 # Test AoC AWS backup operation
-pytest -v -s --ansible-host-pattern=localhost operations/backup -m aoc_aws_backup --junitxml="aoc_aws_backup_results.xml"
+pytest -v -s --ansible-host-pattern=localhost aoc/aws/operations -m aoc_aws_backup --junitxml="aoc_aws_backup_results.xml"
 
 # Test AoC GCP backup operation
-pytest -v -s --ansible-host-pattern=localhost operations/backup -m aoc_gcp_backup --junitxml="aoc_gcp_backup_results.xml"
+pytest -v -s --ansible-host-pattern=localhost aoc/gcp/operations -m aoc_gcp_backup --junitxml="aoc_gcp_backup_results.xml"
 
 # Test AoC AWS restore operation
-pytest -v -s --ansible-host-pattern=localhost operations/restore -m aoc_aws_restore --junitxml="aoc_aws_restore_results.xml"
+pytest -v -s --ansible-host-pattern=localhost aoc/aws/operations -m aoc_aws_restore --junitxml="aoc_aws_restore_results.xml"
 
 # Test AoC GCP restore operation
-pytest -v -s --ansible-host-pattern=localhost operations/backup -m aoc_gcp_restore --junitxml="aoc_gcp_restore_results.xml"
+pytest -v -s --ansible-host-pattern=localhost aoc/gcp/operations -m aoc_gcp_restore --junitxml="aoc_gcp_restore_results.xml"
 ```
