+++ lib/__init__.py
@@ -1,0 +1,5 @@
+"""Lib package.
+
+This package contains additional packages with modules used
+when working with AoC deployments.
+"""
