+++ lib/aoc/__init__.py
@@ -1,0 +1,5 @@
+"""AoC package.
+
+This package contains additional packages with modules used
+when working with AoC deployments.
+"""
