+++ lib/aoc/aws/__init__.py
@@ -1,0 +1,5 @@
+"""AoC aws package.
+
+This package contains additional packages with modules used
+when working with AoC deployments on aws.
+"""
