+++ lib/aoc/aws/operations/__init__.py
@@ -1,0 +1,5 @@
+"""Operations aws package.
+
+This package contains modules performing supported operations for
+AAP deployments on the cloud.
+"""
