+++ tests/aoc/__init__.py
@@ -1,0 +1,4 @@
+"""Aoc tests package.
+
+This package contains test modules validating AoC.
+"""
