+++ tests/aoc/aws/__init__.py
@@ -1,0 +1,5 @@
+"""AoC aws tests package.
+
+This package contains additional packages with modules used
+to test AoC deployments on aws.
+"""
