+++ tests/aoc/aws/operations/__init__.py
@@ -1,0 +1,4 @@
+"""AoC aws operations tests package.
+
+This package contains modules to test AoC on aws backup operations.
+"""
