+++ tests/aoc/gcp/__init__.py
@@ -1,0 +1,5 @@
+"""AoC gcp tests package.
+
+This package contains additional packages with modules used
+to test AoC deployments on gcp.
+"""
