--- tests/lib/__init__.py
@@ -1,5 +1,0 @@
-"""AoC Tests Package.
-
-This package contains additional packages/modules "libraries" that can
-be used by various test suites.
-"""
