--- tests/lib/operations/__init__.py
@@ -1,6 +1,0 @@
-"""Operations Package.
-
-This package contains additional packages/modules "libraries" that
-handle Ansible On Clouds operations using the ops container image.
-"""
-from tests.lib.operations.operations import OperationsBase
