--- tests/lib/operations/backup/__init__.py
@@ -1,5 +1,0 @@
-"""Backup package.
-
-This package contains additional packages/modules "libraries" that
-handle Ansible On Clouds offerings back up operations.
-"""
