--- tests/lib/operations/restore/__init__.py
@@ -1,5 +1,0 @@
-"""Restore package.
-
-This package contains additional packages/modules "libraries" that
-handle Ansible On Clouds offerings restore operations.
-"""
