--- tests/operations/__init__.py
@@ -1,4 +1,0 @@
-"""Operations Test Package.
-
-This package contains various test modules covering AoC operations.
-"""
