--- tests/operations/backup/__init__.py
@@ -1,4 +1,0 @@
-"""Backup operations package.
-
-This package contains test modules validating AoC backups.
-"""
