--- tests/operations/restore/__init__.py
@@ -1,4 +1,0 @@
-"""Restore operations package.
-
-This package contains test modules validating AoC restores.
-"""
